pyproject.toml
[project]
name = "neograph"
version = "0.4.0"
description = "Write Python. Get a production LLM graph. Declarative LLM graph compiler that compiles @node-decorated functions into LangGraph."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Konstantin Mirin", email = "konstantin@mirin.pro"},
]
keywords = ["llm", "langgraph", "declarative", "typed", "pipeline", "graph", "compiler", "ai", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "structlog>=23.0",
    "jsonschema>=4.26.0",
    "json-repair>=0.30",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
langfuse = ["langfuse>=3.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
neograph = "neograph.__main__:main"

[project.urls]
Homepage = "https://neograph.pro"
Documentation = "https://neograph.pro"
Repository = "https://github.com/KonstantinMirin/neograph"
Issues = "https://github.com/KonstantinMirin/neograph/issues"
Changelog = "https://github.com/KonstantinMirin/neograph/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neograph"]

[tool.hatch.build.targets.sdist]
include = [
    "src/neograph",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]
exclude = [
    "tests",
    "examples",
    "website",
    ".claude",
    ".beads",
    "docs",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module-level import not at top (deferred imports for circular dep avoidance)
    "E741",   # ambiguous variable name (n, l are conventional in this codebase)
    "F841",   # local variable assigned but never used (common in side-effect tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["B011"]  # assert False is used in structural guards
"src/neograph/__init__.py" = ["F401"]  # re-exports

[tool.ruff.lint.isort]
known-first-party = ["neograph"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
show_error_codes = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "hypothesis>=6.151.12",
    "pytest-cov>=7.1.0",
]
